cptx: unit test and other fixes
Added unit test for cptx kne cli config push
This commit is dependent on a scrapligo commit hence can't be merged
yet. Contains a couple of comments for the code to be amended post
scrapligo pull request is merged. A new commit will be made on top of
this to mark this PR complete.
It demands changes in scrapligo and a pull request is out on scrapligo.