logshim: new package to send standard error to a log file line by line
This package is inteded to be used when capturing error output from some
other source.  In particular, it can be used with the kne/os/exec package.

This is being added in preperation to switching the logging package from
logrus to klog.  The logrus package provides a function to return what
file descriptor it is logging to.  This doesn't make sense with klog as
klog will rotate files over time.