Add logshim package in preparation for migrating from logrus to klog. (#289)
* Replace examples/3node-withtraffic.pb.txt reference to an existing file.

examples/3node-withtraffic.pb.txt is not in the KNE repository, but
examples/host/3node-host.pb.txt is.

* Fix up more links.  Stop using blob/... references.

* Revert blob references for files with line numbers

* Visit all links and make sure they go someplace

* logshim: new package to send standard error to a log file line by line

This package is inteded to be used when capturing error output from some
other source.  In particular, it can be used with the kne/os/exec package.

This is being added in preperation to switching the logging package from
logrus to klog.  The logrus package provides a function to return what
file descriptor it is logging to.  This doesn't make sense with klog as
klog will rotate files over time.

* Fix the Write call to actually implement io.Writer.

* Fix up typos caught by code review.