Migrate to klog from logrus (#290)
* Replace examples/3node-withtraffic.pb.txt reference to an existing file.

examples/3node-withtraffic.pb.txt is not in the KNE repository, but
examples/host/3node-host.pb.txt is.

* Fix up more links.  Stop using blob/... references.

* Revert blob references for files with line numbers

* Visit all links and make sure they go someplace

* logshim: new package to send standard error to a log file line by line

This package is inteded to be used when capturing error output from some
other source.  In particular, it can be used with the kne/os/exec package.

This is being added in preperation to switching the logging package from
logrus to klog.  The logrus package provides a function to return what
file descriptor it is logging to.  This doesn't make sense with klog as
klog will rotate files over time.

* Fix the Write call to actually implement io.Writer.

* Fix up typos caught by code review.

* Migrate to k8s.io/klog/v2

* Turn off --logtostderr by default

* Remove unused log variables

* Change the logging defaults to both the console and the log file.
The CLI is set up treating logs as the console output so if we
don't also log to stderr there is no output.  To be addressed in
a future CL.

* Clean up comment.
Make sure the default output for execing commands is to be shimmed through logshim.

* Run gofmt

* remove now unused logOut