The new load package reads and does some validation of deployment configs.
The load package has no direct knowledge of the Deployment or
DeploymentConfig structures.

New controllers can be added without needing to modify the kne
deployment code.

The load package can be used independently of the KNE CLI to
validate deployment configurations.

Currently it does not validate that the images are present as this
must be done from within the KNE cluster.