Defer cEOS-lab pod check, update operator version (#534)
* Arista node to implement Status

The Arista node is waiting for its pod to come up in Create(). This is
problematic in high-scale scenarious because we create the pods
sychronously.

Move the check to Status() so checkNodeStatus() handles things instead.

* Update ceoslab operator to v2.0.2 to v2.1.2

See release notes: https://github.com/aristanetworks/arista-ceoslab-operator/releases/tag/v2.1.2

Increased the number of workers in the operator to the number of cores
(from the default value of 1). This may improve performance in
high-scale scenarios. Most of their time is spent generating RSA certificates.